requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
flask
langserve
fastapi
uvicorn
sse_starlette
ollama
openai
langchain-core
langchain_community
huggingface_hub